http_proxy/pyproject.toml
[project]
name = "http-proxy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mitmproxy>=12.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]

[tool.setuptools.packages.find]
include = ["shared*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
